+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - proguard: Added a mandatory `index` field to `JvmFrame` (#1411) by @loewenheim
 
+### Dependencies
+
+- Bump Native SDK from v0.7.0 to v0.7.1 ([#1416](https://github.com/getsentry/symbolicator/pull/1416))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#071)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.7.0...0.7.1)
+
 ## 24.3.0
 
 ### Various fixes & improvements