+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Various fixes & improvements
 
 - proguard: Added a mandatory `index` field to `JvmFrame` (#1411) by @loewenheim
+- proguard: Support for source contex (#1414) by @loewenheim
+- proguard: Field `lineno` on `JvmFrame` is now optional (#1417) by @loewenheim
 
 ## 24.3.0