+++ CHANGELOG.md
@@ -12,9 +12,9 @@
 
 ### Dependencies
 
-- Bump Native SDK from v0.7.0 to v0.7.1 ([#1416](https://github.com/getsentry/symbolicator/pull/1416))
-  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#071)
-  - [diff](https://github.com/getsentry/sentry-native/compare/0.7.0...0.7.1)
+- Bump Native SDK from v0.7.0 to v0.7.2 ([#1416](https://github.com/getsentry/symbolicator/pull/1416), [#1428](https://github.com/getsentry/symbolicator/pull/1428))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#072)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.7.0...0.7.2)
 
 ## 24.3.0